0.2.0.115 ElapsedTime rounding bugfix

+++ ModChecker/Util/Toolkit.cs
@@ -351,7 +351,7 @@ namespace ModChecker.Util
                                            bool showBoth = false,
                                            bool showDecimal = false)
         {
-            double seconds = milliseconds / 1000;
+            double seconds = (double)milliseconds / 1000;
 
             // Decide on when to show seconds, decimals (for seconds only) and minutes
             showBoth = showBoth && threshold > 0 && seconds > threshold;
